Modification de la fonction READER_TPDU_Forge() pour désormais prendre la dataSize sur une valeur de 32 bits (taille max = 256). Ajoit d'un test de vérification sur le paramètre dataSize.

+++ inc/reader_tpdu.h
@@ -80,7 +80,7 @@ READER_Status READER_TPDU_IsNullByte(uint8_t byte);
 READER_Status READER_TPDU_IsSW1(uint8_t byte);
 READER_Status READER_TPDU_IsProcedureByte(uint8_t byte, uint8_t INS);
 
-READER_Status READER_TPDU_Forge(READER_TPDU_Command *tpdu, uint8_t CLA, uint8_t INS, uint8_t P1, uint8_t P2, uint8_t P3, uint8_t *dataBuff, uint8_t dataSize);
+READER_Status READER_TPDU_Forge(READER_TPDU_Command *tpdu, uint8_t CLA, uint8_t INS, uint8_t P1, uint8_t P2, uint8_t P3, uint8_t *dataBuff, uint32_t dataSize);
